+++ flow/project/cli/project.go
@@ -21,7 +21,6 @@ package cli
 import (
 	"errors"
 	"github.com/onflow/flow-cli/flow/project/cli/config/json"
-	"github.com/onflow/flow-cli/flow/project/cli/config/manipulators"
 	"path"
 	"strings"
 
@@ -36,21 +35,21 @@ import (
 
 // Project has all the funcionality to manage project
 type Project struct {
-	composer *manipulators.Composer
+	composer *config.Composer
 	conf     *config.Config
 	accounts []*Account
 }
 
 // LoadProject loads configuration and setup the project
 func LoadProject(configFilePath []string) *Project {
-	composer := manipulators.NewComposer(afero.NewOsFs())
+	composer := config.NewComposer(afero.NewOsFs())
 
 	// here we add all available parsers (more to add yaml etc...)
 	composer.AddConfigParser(json.NewParser())
 	conf, err := composer.Load(configFilePath)
 
 	if err != nil {
-		if errors.Is(err, manipulators.ErrDoesNotExist) {
+		if errors.Is(err, config.ErrDoesNotExist) {
 			Exitf(
 				1,
 				"Project config file %s does not exist. Please initialize first\n",
@@ -74,7 +73,7 @@ func LoadProject(configFilePath []string) *Project {
 
 // ProjectExists checks if project exists
 func ProjectExists(path string) bool {
-	return manipulators.Exists(path)
+	return config.Exists(path)
 }
 
 // InitProject initializes the project
@@ -82,7 +81,7 @@ func InitProject() *Project {
 	emulatorServiceAccount := generateEmulatorServiceAccount()
 
 	return &Project{
-		composer: manipulators.NewComposer(afero.NewOsFs()),
+		composer: config.NewComposer(afero.NewOsFs()),
 		conf:     defaultConfig(emulatorServiceAccount),
 		accounts: []*Account{emulatorServiceAccount},
 	}
@@ -132,7 +131,7 @@ func generateEmulatorServiceAccount() *Account {
 	}
 }
 
-func newProject(conf *config.Config, composer *manipulators.Composer) (*Project, error) {
+func newProject(conf *config.Config, composer *config.Composer) (*Project, error) {
 	accounts, err := accountsFromConfig(conf)
 	if err != nil {
 		return nil, err
